+++ packages/sveltekit/src/server/server-load.ts
@@ -1,35 +1,29 @@
-import type { User, Auth } from "lucia-auth";
 import type { PageData, RequestEvent } from "../types.js";
 import { generateChecksum } from "./crypto.js";
 
 type HandleServerSession = <
-	A extends Auth,
 	LoadFn extends (event: RequestEvent) => any = () => Promise<any>
 >(
-	auth: A,
 	serverLoad?: LoadFn
 ) => (event: RequestEvent) => Promise<Exclude<Awaited<ReturnType<LoadFn>>, void> & PageData>;
 
-export const handleServerSession: HandleServerSession = (auth: Auth, fn?) => {
+export const handleServerSession: HandleServerSession = (fn) => {
 	const handleServerSessionCore = async ({ locals }: RequestEvent): Promise<PageData> => {
-		const session = locals.getSession();
-		try {
-			if (!session) throw new Error();
-			const user = await auth.getUser(session.userId);
+		const { session, user } = await locals.getSessionUser();
+		if (session) {
 			return {
 				_lucia: {
 					user,
 					sessionChecksum: generateChecksum(session.sessionId)
 				}
 			};
-		} catch {
-			return {
-				_lucia: {
-					user: null,
-					sessionChecksum: null
-				}
-			};
 		}
+		return {
+			_lucia: {
+				user: null,
+				sessionChecksum: null
+			}
+		};
 	};
 	return async (event: RequestEvent) => {
 		const { _lucia } = await handleServerSessionCore(event);